Move browser warmup into test support
